Conferindo se já existe aquele email cadastrado

+++ web/tests/support/api/signup/index.js
@@ -9,15 +9,21 @@ export class Api {
 
     async deleteUser(user) {
         const userId = await this.searchUserId(user)
-        const response = await this.request.delete(`${this.baseApi}/usuarios/${userId}`)
-        expect(await response.json()).toEqual(expect.objectContaining({"message": "Registro excluído com sucesso"}));
+        if(userId) {
+            const response = await this.request.delete(`${this.baseApi}/usuarios/${userId}`)
+            expect(await response.json()).toEqual(expect.objectContaining({"message": "Registro excluído com sucesso"}));
+        } else {
+            return "Nenhum registro excluído"
+        }
     }
 
     async searchUserId(user) {
         const userData = await this.request.get(`${this.baseApi}/usuarios?email=${user.email}`)
         expect(await userData.ok()).toBeTruthy()
         const userJson = await userData.json()
-        const userId = userJson.usuarios[0]._id
-        return userId
+        if (userJson.usuarios[0]) {
+            const userId = userJson.usuarios[0]._id
+            return userId
+        }
     }
 }
